move pieces directory under structures directory
