+++ src/lib/handlers/event.handler.ts
@@ -1,7 +1,7 @@
 import { Client, ClientEvents } from 'discord.js';
 import { redBright, yellowBright } from 'colorette';
 import Handler from './handler';
-import Event from '../pieces/event.piece';
+import Event from '../structures/pieces/event.piece';
 import EventStore from '../stores/event.store';
 import Container from '../container';
