+++ src/lib/stores/store.ts
@@ -1,4 +1,4 @@
-import Piece from '../pieces/piece';
+import Piece from '../structures/pieces/piece';
 import PiecesLoader from '../utils/pieces/pieces-loader';
 
 interface Store<T extends Piece> {
